refactor: extrac tag list component and integrated into config list & logs list component
